requirements.txt
django
black
flake8
djangorestframework
python-dotenv
pytest-django
pillow